implemented attaching files to answers
- modified the FilesForPosts model to have an "alias" attribute
to save the file's original name
- refactored and moved some code to attachment.py to be
shared between questions and answers (and later for comments)
- updated dataformat to return the files attached to Posts
- refacted attachService (in questions controller)
- placed a div tag to display links to questions and answers

+++ acj/attachment.py
@@ -1,16 +1,17 @@
 from bouncer.constants import EDIT, CREATE
 from flask import Blueprint, Flask, request
 from flask.ext.login import login_required, current_user
-from flask.ext.restful import Resource
-from acj import db
+from flask.ext.restful import Resource, marshal
+from acj import dataformat, db
 from acj.authorization import allow, require
 from acj.models import PostsForQuestions, FilesForPosts
 from acj.util import new_restful_api
-import os, uuid
+import os, uuid, shutil
 
 attachment_api = Blueprint('attachment_api', __name__)
 api = new_restful_api(attachment_api)
 
+# TODO put in config file
 UPLOAD_FOLDER = os.getcwd() + '/tmpUpload'
 PERMANENT_UPLOAD_FOLDER = os.getcwd() + '/acj/static/pdf'
 ALLOWED_EXTENSIONS = set(['pdf'])
@@ -35,19 +36,42 @@ class AttachmentAPI(Resource):
 		return False
 api.add_resource(AttachmentAPI, '')
 
-# /id
-class AttachmentIdAPI(Resource):
+# /post/post_id
+class AttachmentPostIdAPI(Resource):
 	@login_required
 	def get(self, post_id):
+		# TODO: change to return list of attachments
 		file = FilesForPosts.query.filter_by(posts_id = post_id).first()
 		if file:
-			return {'file': {'id': file.posts_id, 'name': file.name, 'alias': file.alias}}
+			return {'file': marshal(file, dataformat.getFilesForPosts())}
 		return {'file': False}
+api.add_resource(AttachmentPostIdAPI, '/post/<int:post_id>')
+
+# /post/post_id/file_id
+class AttachmentPostIdFileIdAPI(Resource):
 	@login_required
-	def delete(self, post_id):
-		file = FilesForPosts.query.filter_by(posts_id=post_id).first_or_404()
-		tmpName = os.path.join(app.config['PERMANENT_UPLOAD_FOLDER'], file.name)
-		os.remove(tmpName)
-		db.session.delete(file)
-		db.session.commit()
-api.add_resource(AttachmentIdAPI, '/<int:post_id>')
+	def delete(self, post_id, file_id):
+		file = FilesForPosts.query.filter_by(posts_id=post_id).filter_by(id=file_id).first()
+		if file:
+			tmpName = os.path.join(app.config['PERMANENT_UPLOAD_FOLDER'], file.name)
+			os.remove(tmpName)
+			db.session.delete(file)
+			db.session.commit()
+
+api.add_resource(AttachmentPostIdFileIdAPI, '/post/<int:post_id>/<int:file_id>')
+
+# move file from temporary location to "permanent" location
+# makes a FilesForPosts entry
+def addNewFile(alias, name, course_id, question_id, post_id):
+	tmpName = str(course_id) + '_' + str(question_id) + '_' + str(post_id) + '.pdf'
+	shutil.move(os.path.join(app.config['UPLOAD_FOLDER'], name), os.path.join(app.config['PERMANENT_UPLOAD_FOLDER'], tmpName))
+	file = FilesForPosts(posts_id=post_id, author_id=current_user.id, name=tmpName, alias=alias)
+	db.session.add(file)
+	db.session.commit()
+
+# delete file from Post
+def deleteFile(post_id):
+	# TODO: delete ALL files under post with post_id
+	file = FilesForPosts.query.filter_by(posts_id=post_id).first()
+	if file:
+		os.remove.remove(app.config['PERMANENT_UPLOAD_FOLDER'], file.name)
